Remove "-o default_permissions" unless needed.
It is only needed when "-o allow_other" is specified.

"-o default_permissions" causes libfuse to check file access
in userspace. This costs CPU cycles and causes additional
stat() calls - libfuse has to walk up the whole path to check
for "x" permissions on directories.

This improves "make benchmark-reverse" performance
by 30% when caching is disabled. It also gives a slight
improvement with caches on.

Before:

  tests/benchmark-reverse.pl /var/tmp
  * rsync 1 (initial copy)... 12179 ms
  * rsync 2 (no changes)... 1840 ms
  cleaning up... done
  tests/benchmark-reverse.pl /var/tmp --nocache
  * rsync 1 (initial copy)... 30696 ms
  * rsync 2 (no changes)... 10552 ms
  cleaning up... done

After:

  tests/benchmark-reverse.pl /var/tmp
  * rsync 1 (initial copy)... 12095 ms
  * rsync 2 (no changes)... 1693 ms
  cleaning up... done
  tests/benchmark-reverse.pl /var/tmp --nocache
  * rsync 1 (initial copy)... 21266 ms
  * rsync 2 (no changes)... 6486 ms
  cleaning up... done