Initialize dummy variable
Building fails with a warning that we didn't init this var.  So I'm
setting it to 0.  Is that right, wrong, catastrophic?  I dunno, but
the warning went away and the build succeeds.  Maybe it would be
better to just -Wno-error=maybe-uninitialized in
vendor/github.com/google/googletest/googletest/cmake/internal_utils.cmake
or something.